Refactor: 멘토링 당시 받은 리뷰
