+++ pages/myShopInfo.tsx
@@ -1,3 +1,4 @@
+import { useMediaQuery } from "react-responsive";
 import Button from "@/shared/@common/ui/Button/Button";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
@@ -11,12 +12,13 @@ const myShopInfo = () => {
   }
 
   const [size, setSize] = useState('large');
+  const isMobile = useMediaQuery({ query: '(max-width: 767px)' });
 
   useEffect(() => {
-    const handleSize = () => {
-      if(window.innerWidth < 768){
-        setSize('small');
-      }
+    if (isMobile) {
+      setSize('small');
+    } else {
+      setSize('large');
     }
   }, []);
 
@@ -28,7 +30,9 @@ const myShopInfo = () => {
         <p className="text-black text-[28px] font-bold">내 가게</p>
         <div className="flex w-full py-[60px] px-[24px] flex-col content-center items-center gap-6 rounded-xl border border-solid border-gray-20">
           <p className=" self-stretch text-center text-base">내 가게를 소개하고 공고도 등록해 보세요.</p>
-          <Button size={size} color='colored' onClick={handleWritingShopInfo} content='가게 등록하기' />
+          <Button size={size} color='colored' onClick={handleWritingShopInfo} >
+            가게 등록하기
+          </Button>
         </div>
       </div>
     </>
